feat: Add memory benchmarking to spammer
