+++ Cargo.toml
@@ -22,6 +22,8 @@ secp256k1 = { version = "0.20.3", features = ["global-context", "serde"] }
 sha2 = "0.9.5"
 base58 = "0.1.0"
 hex = "0.4.3"
+jemallocator = "0.1.9"
+jemalloc-ctl = "0.2.0"
 lazy_static = "1.4.0"
 rand = "0.8.4"
 warp = "0.3"
