+++ src/bin/spammer.rs
@@ -1,5 +1,6 @@
 use rayon::iter::IntoParallelIterator;
 use rayon::prelude::*;
+use saito_rust::test_utilities::memory_stats::MemoryStats;
 use saito_rust::{
     block::{Block, BlockCore, TREASURY},
     blockchain::AddBlockEvent,
@@ -16,7 +17,7 @@ pub async fn main() -> saito_rust::Result<()> {
     let keypair = Keypair::new();
 
     let (mut blockchain, mut slips) =
-        test_utilities::make_mock_blockchain_and_slips(&keypair, 3 * 100000).await;
+        test_utilities::mocks::make_mock_blockchain_and_slips(&keypair, 3 * 100000).await;
     let prev_block = blockchain.latest_block().unwrap();
 
     let mut prev_block_hash = prev_block.hash().clone();
@@ -78,6 +79,8 @@ pub async fn main() -> saito_rust::Result<()> {
 
         start_ts = create_timestamp();
         println!("ADD BLOCK {}", block.id());
+        println!("Memory Stats {}", MemoryStats::current());
+
         let result = blockchain.add_block(block).await;
         assert!(result == AddBlockEvent::AcceptedAsLongestChain);
         finish_ts = create_timestamp();
